+++ src/MPLEconddeg.h
@@ -23,12 +23,12 @@ void MPLEconddeg_wrapper (int *heads, int *tails, int *dnedges,
                    int *mheads, int *mtails, int *mdnedges);
 void CondDegSampler (char *MHproposaltype, char *MHproposalpackage,
 		 double *theta, double *networkstatistics, 
-		 long int samplesize, long int burnin, 
-		 long int interval, int hammingterm, int fVerbose,
+		 int samplesize, int burnin, 
+		 int interval, int hammingterm, int fVerbose,
 	       	 Network *nwp, Model *m, DegreeBound *bd);
 void CondDegSample (MHproposal *MHp,
 			 double *theta, double *statistics, 
-			 long int nsteps, long int *staken,
+			 int nsteps, int *staken,
 			 int hammingterm, int fVerbose,
 			 Network *nwp, Model *m, DegreeBound *bd);
 #endif
